requirements.txt
flask==3.0.3
requests==2.32.3
pymongo==4.8.0
Flask-Cors==4.0.1
websocket-client==1.8.0